Crate relm4_components

Collection of reusable and easily configurable components for Relm4.

Docs of related crates: relm4 | relm4-macros | relm4-components | relm4-css | gtk4-rs | gtk-rs-core | libadwaita-rs | libpanel-rs

GitHub | Website | Book | Blog

Modules

alert

Reusable and easily configurable alert component.

open_button

Reusable and easily configurable open button dialog component.

open_dialog

Reusable and easily configurable open dialog component.

save_dialog

Reusable and easily configurable save dialog component.

simple_adw_combo_rowlibadwaita

A wrapper around adw::ComboRow that makes it easier to use from regular Rust code.

simple_combo_box

A wrapper around gtk::ComboBoxText that makes it easier to use from regular Rust code.

web_imageweb

Reusable and easily configurable component for loading images from the web.