Trait relm4_components::alert::AlertParent

pub trait AlertParent: Model where
    Self::Widgets: AlertParentWidgets, {
    fn alert_config(&self) -> AlertSettings;

    fn confirm_msg() -> Self::Msg;

    fn cancel_msg() -> Self::Msg;

    fn option_msg() -> Self::Msg;
}

Interface for the parent model

Required methods

fn alert_config(&self) -> AlertSettings

Configuration for alert component.

fn confirm_msg() -> Self::Msg

Message sent to parent if user clicks confirm button

fn cancel_msg() -> Self::Msg

Message sent to parent if user clicks cancel button

fn option_msg() -> Self::Msg

Message sent to parent if user clicks third option button