relm4_components::open_button

Struct OpenButtonSettings

pub struct OpenButtonSettings {
    pub dialog_settings: OpenDialogSettings,
    pub icon: Option<&'static str>,
    pub text: &'static str,
    pub recently_opened_files: Option<&'static str>,
    pub max_recent_files: usize,
}

Configuration for the open button component

Fields
dialog_settings: OpenDialogSettings

Settings for the open file dialog.

icon: Option<&'static str>

Icon of the open button.

text: &'static str

Text of the open button.

recently_opened_files: Option<&'static str>

Path to a file where recent files should be stored. This list is updated fully automatically.

max_recent_files: usize

Maximum amount of recent files to store. This is only used if a path for storing the recently opened files was set.