relm4_components::open_dialog

Type Alias OpenDialogMulti

pub type OpenDialogMulti = OpenDialogInner<MultiSelection>;

A component that prompts the user to choose a file.

The user would be able to select multiple files. If you’d like to select just one, use OpenDialog.

Aliased Type

pub struct OpenDialogMulti { /* private fields */ }