Module relm4_components::save_dialog

Reusable and easily configurable save dialog component.

Example implementation

Structs

SaveDialogModel

Model of the save dialog component

SaveDialogSettings

Configuration for the save dialog component

SaveDialogWidgets

Widgets for the save dialog

Enums

SaveDialogMsg

Messages that can be sent to the save dialog component

Traits

SaveDialogConfig

Interface for building the configuration for SaveDialog

SaveDialogParent

Interface for the parent model of the save dialog