Struct relm4_components::save_dialog::SaveDialogModel

pub struct SaveDialogModel<Conf: SaveDialogConfig> { /* private fields */ }

Model of the save dialog component

Implementations

impl<Conf: SaveDialogConfig> SaveDialogModel<Conf>

pub fn track_all() -> u8

Use this to check whether any changes made to this struct.

pub fn mark_all_changed(&mut self)

Use this to mark all fields of the struct as changed.

pub fn changed(&self, mask: u8) -> bool

Check for changes made to this struct.

pub fn reset(&mut self)

Resets the tracker of this struct.