Crate relm4_components

Collection of reusable and easily configurable components for Relm4.

The docs are available in two versions. Use the stable docs if you want get information about a version that was already published. Visit the nightly docs if are trying out the newest but possibly unstable version of the crate.

Docs of related crates: relm4 | relm4-macros | relm4-components | gtk4-rs | gtk-rs-core | libadwaita-rs | libpanel-rs

GitHub | Website | Book | Blog

Modules

alert

Reusable and easily configurable alert component.

open_button

Reusable and easily configurable open button dialog component.

open_dialog

Reusable and easily configurable open dialog component.

save_dialog

Reusable and easily configurable save dialog component.

simple_combo_box

A wrapper around gtk::ComboBoxText that makes it easier to use from regular Rust code.

web_image

Reusable and easily configurable component for loading images from the web.