Module relm4_components::open_dialog

Reusable and easily configurable open dialog component.

Example implementation

Structs

MultiSelection

A type of selection where multiple types can be chosen at a time.

OpenDialogInner

Model for the open dialog component

OpenDialogInnerWidgets

Widgets of the open dialog component.

OpenDialogSettings

Configuration for the open dialog component

SingleSelection

A type of selection where only one file can be chosen at a time.

Enums

OpenDialogMsg

Messages that can be sent to the open dialog component

OpenDialogResponse

Messages that can be sent from the open dialog component

Traits

Select

Type of selection used for the open dialog.

Type Definitions

OpenDialog

A component that prompts the user to choose a file.

OpenDialogMulti

A component that prompts the user to choose a file.